Move settings refreshing to App
